% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_ancestry_functions.R
\name{get_la_example}
\alias{get_la_example}
\title{Get local population-specific PRS in a window of determined start and end}
\usage{
get_la_example(la_df, window_size)
}
\arguments{
\item{bigSNP.obj}{\itemize{
\item bigsnp object containing genotypes, window_size - size of blocks/windows
}}
}
\value{
dataframe containing local ancestry inference in windows
}
\description{
This function matches summary stats with loaded plink files using bigsnpr::snp_matchs. Matches alleles and flips if needed.
}
