% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_local_PRS_functions.R
\name{get_local_PRS}
\alias{get_local_PRS}
\title{Get local population-specific PRS in a window of determined start and end}
\usage{
get_local_PRS(
  bigSNP.obj,
  ind.idx,
  start,
  end,
  chrom,
  r2,
  thres,
  matched_betas_pvals
)
}
\arguments{
\item{bigSNP.obj}{\itemize{
\item bigsnp object containing genotypes, window_size - size of blocks/windows
}}
}
\value{
dataframe containing local population-specific PRS in regions of window_size
}
\description{
This function matches summary stats with loaded plink files using bigsnpr::snp_matchs. Matches alleles and flips if needed.
}
