% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_plinkFile}
\alias{load_plinkFile}
\title{Load plink file using bigSNP package}
\usage{
load_plinkFile(inFile)
}
\arguments{
\item{inFile}{Path to plink bed file including .bed extension}
}
\value{
A matrix of the infile
}
\description{
This function loads a plink file into R. Uses bigSNP package to first check if .rds file exists. If not will create from .bed file and then import
}
