% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_local_PRS_functions.R
\name{make_local_PRS_df}
\alias{make_local_PRS_df}
\title{Make dataframe containing local population-specific PRS in each window of determined size looping through entire chromosome file}
\usage{
make_local_PRS_df(
  bigSNP.obj,
  matched_betas_pvals,
  window_size_num,
  r_sq,
  thres,
  out_file_path
)
}
\arguments{
\item{bigSNP.obj}{\itemize{
\item bigsnp object containing genotypes, window_size - size of blocks/windows
}}
}
\value{
dataframe containing local population-specific PRS in regions of window_size
}
\description{
Make dataframe containing local population-specific PRS in each window of determined size looping through entire chromosome file
}
